refactor(workspace): merge standard-schema/ directory into single file
Combine types.ts, to-json-schema.ts, and index.ts into a single
standard-schema.ts file. Update all internal imports. Delete the
directory and README. No public API change — same exports, single file.
