--- packages/workspace/src/shared/standard-schema/index.ts
@@ -1,2 +1,0 @@
-export { standardSchemaToJsonSchema } from './to-json-schema.js';
-export type { CombinedStandardSchema } from './types.js';
