+++ packages/workspace/src/workspace/describe-workspace.ts
@@ -27,7 +27,7 @@
 import type { StandardJSONSchemaV1 } from '@standard-schema/spec';
 import type { TSchema } from 'typebox';
 import { iterateActions } from '../shared/actions.js';
-import { standardSchemaToJsonSchema } from '../shared/standard-schema/to-json-schema.js';
+import { standardSchemaToJsonSchema } from '../shared/standard-schema.js';
 import type { AnyWorkspaceClient } from './types.js';
 
 // ════════════════════════════════════════════════════════════════════════════
